Tratamento de exceções

+++ src/main/java/com/dtlsilva/mhglavacar/services/ServicoService.java
@@ -1,17 +1,22 @@
 package com.dtlsilva.mhglavacar.services;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.dtlsilva.mhglavacar.dto.ServicoDTO;
 import com.dtlsilva.mhglavacar.entidades.servicos.Servico;
 import com.dtlsilva.mhglavacar.repositories.ServicoRepository;
+import com.dtlsilva.mhglavacar.services.exceptions.DatabaseException;
+import com.dtlsilva.mhglavacar.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ServicoService {
@@ -27,7 +32,8 @@ public class ServicoService {
 	@Transactional(readOnly = true)
 	public ServicoDTO findById(Long id) {
 		Optional<Servico> result = repository.findById(id);
-		Servico servico = result.get();
+		Servico servico = result.orElseThrow(
+				() -> new ResourceNotFoundException("Serviço não encontrado"));
 		ServicoDTO dto = new ServicoDTO(servico);
 		return dto;
 	}
@@ -42,15 +48,28 @@ public class ServicoService {
 	
 	@Transactional
 	public ServicoDTO update(Long id, ServicoDTO dto) {
-		Servico entity = repository.getReferenceById(id);
-		dtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ServicoDTO(entity);
+		try {
+			Servico entity = repository.getReferenceById(id);
+			dtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ServicoDTO(entity);
+		}
+		catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Serviço não encontrado");
+		}
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		try {
+			if (!repository.existsById(id)) {
+				throw new ResourceNotFoundException("Serviço não encontrado");
+			}
+			repository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DatabaseException("Falha de integridade referencial");
+		}
 	}
 	
 	private void dtoToEntity(ServicoDTO dto, Servico entity) {
